Remove unnecessary isolated memory pages in igvmfile.py
The generated IGVMfile includes bzImage code and some memory used by
startup_32 as isolated pages, and leaves invalidated gPA as holes in
e820 table. The guest startup_64 will pvalidate the gPAs on demand if
the needed memory is not inside the e820 table.

This patch reduces #launch_update from ~8k to ~1k.