UBUNTU: SAUCE: Drivers: hv: vmbus: Remove second mapping of VMBus monitor pages
https://github.com/kelleymh/linux.git 1b00968e93e56a83a6de7f720ed443372d6d8efc

With changes to how Hyper-V guest VMs flip memory between private
(encrypted) and shared (decrypted), creating a second kernel virtual
mapping for shared memory is no longer necessary.  Everything needed
for the transition to shared is handled by set_memory_decrypted().

As such, remove the code to create and manage the second
mapping for VMBus monitor pages. Because set_memory_decrypted()
and set_memory_encrypted() are no-ops in normal VMs, it's
not even necessary to test for being in a Confidential VM
(a.k.a., "Isolation VM").

Signed-off-by: Tim Gardner <tim.gardner@canonical.com>